tms_master/requirements.txt
django
psycopg2-binary
djangorestframework
django-cors-headers
docker